+++ bacteria.h
@@ -9,33 +9,34 @@
 class Water;
 class Genome;
 struct Command{
-    int cmd;
-    static constexpr int offset_conut = 3;
-    int offset[offset_conut];
+    uint8_t cmd;
+    //uint8_t param;
+    static constexpr uint8_t offset_conut = 3;
+    int16_t offset[offset_conut];
     static constexpr int command_count = 126;
-    int offset_mutate_chance_10000;
-    int cmd_mutate_chance_10000;
+    uint16_t offset_mutate_chance_10000;
+    uint16_t cmd_mutate_chance_10000;
 };
 class Bacteria
 {
 protected:
     Water* water;
     std::shared_ptr<Genome> genome;
-    int x,y;
-    int energy;
-    int attack;
+    uint16_t x,y;
+    int32_t energy;
+    uint16_t attack;
     const Bacteria* parent;
-    int birth_time;
-    int cmd_ptr = 0;
-    int COUNT_OF_ACTION = 0;
+    uint64_t birth_time;
+    int16_t cmd_ptr = 0;
+    uint8_t COUNT_OF_ACTION = 0;
     bool stop_action;
     bool killed = false;
-    int killer = 0;
-    int atp = 0;
+    uint8_t killer = 0;
+    uint32_t atp = 0;
 public:
-    Bacteria(Water* water, int x, int y,int energy,const Bacteria* parent);
+    Bacteria(Water* water, uint16_t x, uint16_t y,int32_t energy,const Bacteria* parent);
     void Tick();
-    void TryMove(int dx,int dy);
+    void TryMove(int8_t dx,int8_t dy);
     void Kill(bool in_cycle=false);
     void Clone();
     void Photosynthesis();
@@ -53,12 +54,12 @@ public:
     void cmd_HowMuchEnergy(Command cmd);
 
     std::shared_ptr<Genome> GetGenome() const;
-    void GetCoords(int &X,int &Y);
-    int GetEnergy();
-    int GetBirthTime();
-    int GetAttack();
-    void IncreaseEnergy(int added_energy);
-    int GetKiller();
+    void GetCoords(uint16_t &X,uint16_t &Y);
+    int32_t GetEnergy();
+    uint64_t GetBirthTime();
+    uint16_t GetAttack();
+    void IncreaseEnergy(int32_t added_energy);
+    uint8_t GetKiller();
     bool GetKilled();
 };
